Refactor: Rename .js files to .jsx for better rendering
- Changed file extensions from .js to .jsx for improved browser rendering.
- Reformatted files for better code readability.
- Updated import statements to ensure proper rendering with new .jsx extensions.
