+++ src/front/js/component/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import logo from "../../img/logo.png";
-import { Dropdown } from "./Dropdown.jsx";
+import { Dropdown } from "./Dropdown";
 import { ThemeSwitcher } from "./ThemeSwitcher";
 import "../../styles/index.css";
