+++ src/front/js/pages/favorite.jsx
