+++ src/front/js/pages/single.jsx
