+++ src/front/js/pages/suggestions.jsx
