+++ src/front/js/store/appContext.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import getState from "./flux.js";
+import getState from "./flux.jsx";
 
 export const Context = React.createContext(null);
