+++ webpack.common.js
@@ -5,7 +5,7 @@ const Dotenv = require('dotenv-webpack');
 
 module.exports = {
   entry: [
-    './src/front/js/index.js'
+    './src/front/js/index.jsx'
   ],
   output: {
     filename: 'bundle.js',
